agroflow-pro/requirements.txt
# AgroFlow Pro - Python Dependencies

# Web Framework
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5

# Database
SQLAlchemy==2.0.19
psycopg2-binary==2.9.6
pymongo==4.3.3

# IoT & Real-time Communication
paho-mqtt==1.6.1
python-socketio==5.9.0
python-engineio==4.5.1

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# APIs & Webhooks
requests==2.31.0
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
python-jose==3.3.0

# Cloud Integration
boto3==1.28.0  # AWS SDK
google-cloud-storage==2.10.0

# SMS & Notifications
twilio==8.7.0  # For SMS alerts

# Logging & Monitoring
python-logging==2.1
sentry-sdk==1.28.1

# Development
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Data Visualization
matplotlib==3.7.2
plotly==5.15.0

# Documentation
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0

# Task Scheduling
celery==5.3.1
redis==4.6.0
